src/test/java/Day08Tests.java
import com.olsson.aoc2021.Day08;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

class Day08Tests {

    @Test
    void part1Example() {
        var day = new Day08();
        assertEquals("26", day.part1(example.lines().toList()));
    }

    @Test
    void part1Solution() {
        var day = new Day08();
        assertEquals("274", day.part1(day.getInput()));
    }

    @Test
    void part2ExampleSmall() {

        var day = new Day08();
        assertEquals("5353", day.part2(example2.lines().toList()));
    }

    @Test
    void part2Example() {
        var day = new Day08();
        assertEquals("61229", day.part2(example.lines().toList()));
    }

    @Test
    void part2Solution() {
        var day = new Day08();
        assertEquals("61229", day.part2(day.getInput()));
    }

    String example = """
            be cfbegad cbdgef fgaecd cgeb fdcge agebfd fecdb fabcd edb | fdgacbe cefdb cefbgd gcbe
            edbfga begcd cbg gc gcadebf fbgde acbgfd abcde gfcbed gfec | fcgedb cgb dgebacf gc
            fgaebd cg bdaec gdafb agbcfd gdcbef bgcad gfac gcb cdgabef | cg cg fdcagb cbg
            fbegcd cbd adcefb dageb afcb bc aefdc ecdab fgdeca fcdbega | efabcd cedba gadfec cb
            aecbfdg fbg gf bafeg dbefa fcge gcbea fcaegb dgceab fcbdga | gecf egdcabf bgf bfgea
            fgeab ca afcebg bdacfeg cfaedg gcfdb baec bfadeg bafgc acf | gebdcfa ecba ca fadegcb
            dbcfg fgd bdegcaf fgec aegbdf ecdfab fbedc dacgb gdcebf gf | cefg dcbef fcge gbcadfe
            bdfegc cbegaf gecbf dfcage bdacg ed bedf ced adcbefg gebcd | ed bcgafe cdgba cbgef
            egadfb cdbfeg cegd fecab cgb gbdefca cg fgcdab egfdb bfceg | gbdfcae bgc cg cgb
            gcafb gcf dcaebfg ecagb gf abcdeg gaef cafbge fdbac fegbdc | fgae cfgab fg bagce
            """;

    String example2 = "acedgfb cdfbe gcdfa fbcad dab cefabd cdfgeb eafb cagedb ab | cdfeb fcadb cdfeb cdbaf";

}
